made some changes to the telop and the auto

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoverCollector.java
@@ -87,7 +87,7 @@ public class RoverCollector {
 //    }
 
     public void setPositionDegrees(double targetDegrees) {
-        setPosition((int)convertDegreesToEncoder(targetDegrees));
+        setPosition(convertDegreesToEncoder(targetDegrees));
     }
 
     public void setPosition(int targetEncoder) {
